IGNITE-16505 Move row count approximation method from ignite-indexing
