+++ modules/indexing/src/main/java/org/apache/ignite/internal/processors/query/h2/opt/GridH2Table.java
@@ -25,7 +25,6 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicIntegerFieldUpdater;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.LongAdder;
@@ -116,12 +115,6 @@ public class GridH2Table extends TableBase {
      * table size change exceeds this threshold. Should be the number in interval (0,1).
      */
     private static final double STATS_UPDATE_THRESHOLD = 0.1; // 10%.
-
-    /** */
-    private static final int STATS_CLI_UPDATE_THRESHOLD = 200;
-
-    /** */
-    AtomicInteger cliReqCnt = new AtomicInteger();
 
     /** Cache context info. */
     private final GridCacheContextInfo cacheInfo;
@@ -1164,13 +1157,6 @@ public class GridH2Table extends TableBase {
         return tblStats.primaryRowCount();
     }
 
-    /** */
-    public long getRowCountApproximationNoCheck() {
-        refreshStatsIfNeededEx();
-
-        return tblStats.primaryRowCount();
-    }
-
     /**
      * @param qctx Context.
      *
@@ -1200,26 +1186,6 @@ public class GridH2Table extends TableBase {
             size.add(totalRowCnt);
 
             tblStats = new TableStatistics(totalRowCnt, primaryRowCnt);
-        }
-    }
-
-    /**
-     * Refreshes table stats if they are possibly outdated, must be called only in client mode.
-     */
-    private void refreshStatsIfNeededEx() {
-        if (cliReqCnt.getAndIncrement() % STATS_CLI_UPDATE_THRESHOLD == 0) {
-            TableStatistics stats = tblStats;
-
-            long primaryRowCnt = stats.primaryRowCount();
-
-            try {
-                primaryRowCnt = cacheInfo.cacheContext().cache().size(new CachePeekMode[] {CachePeekMode.PRIMARY});
-            }
-            catch (IgniteCheckedException e) {
-                log.warning("Can`t update cache size.", e);
-            }
-
-            tblStats = new TableStatistics(stats.totalRowCount(), primaryRowCnt);
         }
     }
